Conflicts:
	src/multi.c

+++ src/utils.h
@@ -1,5 +1,5 @@
 #include "empr.h"
 #include "ir_sensor.h"
 #include "ultrasound.h"
-void utils_process_digit_input(int last_key_press, uint16_t* result);
+void utils_process_digit_input(int last_key_press, uint32_t* result);
 uint32_t utils_get_ir_and_ultrasound_distance();
